Trait pallet_root_offences::pallet::Config

pub trait Config:
    Config
    + Config
    + Config<ValidatorId = <Self as Config>::AccountId>
    + Config<FullIdentification = Exposure<<Self as Config>::AccountId, BalanceOf<Self>>, FullIdentificationOf = ExposureOf<Self>> {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Object Safety

This trait is not object safe.